Fixes #15932 - Stricter check on file globs
When creating a certificate named foreman-proxy-client and then one for
foreman-proxy meant that the glob would match the client. This meant it
created foreman-proxy-client-1.0-1 and then foreman-proxy-1.0-2 even
though there never was a foreman-proxy-1.0-1.

By matching on a number, this can still happen if the user requested a
certificate where the package name included a number. However, this is
currently not the case in Satellite which is the only known consumer of
this package.

A better version would actually use the name from the RPM.